// WEEK2/httpClient/src/app/models/Product.ts
export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}

export interface ProductViewModel {
  id: number,
  title: string,
  category: string,
  price: number
}

export interface ProductResponse {
  products: Product[]
}